Arreglo banana y granada

+++ server/items/bullets/granada_bullet.h
@@ -15,9 +15,8 @@ private:
   bool continue_moving;
 
 public:
-  GranadaBullet(uint8_t type, uint8_t id, float x_pos, float y_pos,
-                uint8_t damage, uint8_t range, float spread,
-                uint8_t time_to_explode_);
+  GranadaBullet(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
+                uint8_t range, float spread, uint8_t time_to_explode_);
   void executeAction() override;
   void release_granada(float x_pos, float y_pos, uint8_t direction,
                        uint8_t time_to_explode);
