+++ server/items/weapons/ak47.cpp
@@ -8,7 +8,7 @@ constexpr int MAX_SPREAD_COUNTER = -1;
 
 constexpr int BURST_INTERVAL = 5;
 
-Ak47::Ak47(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+Ak47::Ak47(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
            uint8_t range, int ammo_quantity, float recoil)
     : Weapon(type, id, x_pos, y_pos, damage, range, ammo_quantity, recoil),
       spread_counter(0), bullets_vector() {
