+++ server/items/weapons/banana.h
@@ -14,7 +14,7 @@ private:
   uint8_t time_to_explode;
 
 public:
-  Banana(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+  Banana(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
          uint8_t range, int ammo_quantity, float recoil);
   bool isEmptyAmmo() override;
   bool isActive() override;
