+++ server/items/weapons/granada.h
@@ -14,7 +14,7 @@ private:
   uint8_t time_to_explode;
 
 public:
-  Granada(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+  Granada(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
           uint8_t range, int ammo_quantity, float recoil);
   bool isEmptyAmmo() override;
   std::unique_ptr<Bullet> shoot(bool is_aiming_up) override;
