+++ server/items/weapons/laser_rifle.h
@@ -13,7 +13,7 @@ private:
   std::vector<Bullet> bullets_vector;
 
 public:
-  LaserRifle(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+  LaserRifle(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
              uint8_t range, int ammo_quantity, float recoil);
 
   bool isEmptyAmmo() override;
