+++ server/items/weapons/magnum.h
@@ -9,7 +9,7 @@ private:
   Bullet bala;
 
 public:
-  Magnum(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+  Magnum(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
          uint8_t range, int ammo_quantity, float recoil);
 
   bool isEmptyAmmo() override;
