+++ server/items/weapons/pewpew_laser.h
@@ -14,8 +14,8 @@ private:
   int beam_count;
 
 public:
-  PewPewLaser(uint8_t type, uint8_t id, float x_pos, float y_pos,
-              uint8_t damage, uint8_t range, int ammo_quantity, float recoil);
+  PewPewLaser(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
+              uint8_t range, int ammo_quantity, float recoil);
 
   bool isEmptyAmmo() override;
