+++ server/items/weapons/sniper.h
@@ -12,7 +12,7 @@ private:
   Bullet bala;
 
 public:
-  Sniper(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
+  Sniper(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
          uint8_t range, int ammo_quantity, float recoil);
 
   bool isEmptyAmmo() override;
